7.1.1 разбиваем редьюсер
